requirements.txt
torch>=2.0.0
torchvision>=0.15.0
gradio>=3.40
Pillow>=10.0.0
timm>=0.9.0
